libstagefright: Small buffer hint for PCM offload
Audioplayer when explicitly requesting PCM offload
needs to inform audio track that small buffers have
to be used. This is required to make sure that pcm
offloaded tracks work very similar to deep buffer
path. This hint is passed to the HAL which makes
the write calls are blocking. Handle the new
properties for pcm 16 and 24 bit offload in policy
manager

CRs-fixed: 749148
Change-Id: Ia7b2260ec6e4821a7dbbfcc3ec7719c334aa40f6